Move timeline tool code to a separate folder
