+++ static/js/tools/statvar_menu/menu.tsx
@@ -16,9 +16,9 @@
 
 import React, { Component } from "react";
 import axios from "axios";
-import { StatsVarNode } from "./timeline_util";
-import hierarchy from "../../data/hierarchy_top.json";
-import { StatsVarFilterInterface } from "./commons";
+import { StatsVarNode } from "../timeline/util";
+import hierarchy from "../../../data/hierarchy_top.json";
+import { StatsVarFilterInterface } from "../commons";
 import _ from "lodash";
 
 const jsonPath = "../../data/hierarchy_statsvar.json";
