+++ static/js/tools/timeline/info.tsx
