+++ static/js/tools/timeline/page.test.tsx
@@ -15,14 +15,14 @@
  */
 
 jest.mock("axios");
-jest.mock("../chart/draw");
+jest.mock("../../chart/draw");
 import Enzyme, { mount } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 import React from "react";
 import axios from "axios";
 import pretty from "pretty";
-import { Page } from "./timeline_page";
-import { axios_mock, drawGroupLineChart_mock } from "./mock_functions";
+import { Page } from "./page";
+import { axios_mock, drawGroupLineChart_mock } from "../mock_functions";
 
 Enzyme.configure({ adapter: new Adapter() });
