+++ static/js/tools/timeline/timeline.ts
@@ -20,7 +20,7 @@
 
 import React from "react";
 import ReactDOM from "react-dom";
-import { Page } from "./timeline_page";
+import { Page } from "./page";
 
 window.onload = () => {
   ReactDOM.render(
