+++ static/js/tools/timeline/util.test.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import { getPlaceNames, TimelineParams } from "./timeline_util";
+import { getPlaceNames, TimelineParams } from "./util";
 
 test("get place names", () => {
   const dcids = ["geoId/4459000", "country/USA"];
